fix: Feedback from nikolaia
