+++ src/CvPartner/CVPartnerRepository.cs
@@ -1,6 +1,27 @@
-﻿namespace CvPartner;
+﻿using CvPartner.DTOs;
+using CvPartner.Utils;
+
+using Microsoft.Extensions.Options;
+
+using Refit;
+
+using Shared;
+namespace CvPartner;
 
 public class CVPartnerRepository
 {
-    // TODO: Use Refit and Access Token to call CVPartner
+    private readonly IOptionsSnapshot<AppSettings> _appSettings;
+
+    public CVPartnerRepository(IOptionsSnapshot<AppSettings> appSettings)
+    {
+        _appSettings = appSettings;
+    }
+
+    public async Task<IEnumerable<CVPartnerUserDTO>> GetCVPartnerDTO()
+    {
+        var cvPartnerApi = RestService.For<Interfaces.IEmployeeApi>("https://variant.cvpartner.com/api/v1");
+
+        var employees = await cvPartnerApi.GetAllEmployee(_appSettings.Value.Token);
+        return employees;
+    }
 }
