+++ src/WebApi/Program.cs
@@ -2,8 +2,6 @@ using Azure.Identity;
 using Employee.Repositories;
 using Microsoft.Azure.Services.AppAuthentication;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration.Ini;
-
 using Shared;
 using Shared.AzureIdentity;
